backend/requirements.txt
Flask
spacy
pdfminer.six
flask-cors
python-docx
openai>=1.0.0
python-dotenv>=1.0.0